samples/python/redmist-sample/pyproject.toml
[tool.poetry]
name = "redmist-sample"
version = "0.1.0"
description = "A sample RedMist console application"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
msgpack = "^1.0.0"
signalrcore = "^0.9.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
